requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==3.2.18
django-debug-toolbar==4.0.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
Faker==18.4.0
faker-vehicle==0.2.0
filelock==3.12.0
importlib-metadata==6.5.0
inflection==0.5.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.22.0
Werkzeug==2.2.3
zipp==3.15.0
